+++ Yuta-KTD/flutter_challenge1_yuta_ktd/.gitignore
@@ -42,3 +42,6 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+
+# env
+.env
