+++ Yuta-KTD/flutter_challenge1_yuta_ktd/analysis_options.yaml
@@ -23,7 +23,7 @@ linter:
   # producing the lint.
   rules:
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
-    # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+    prefer_single_quotes: true 
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
