+++ Yuta-KTD/flutter_challenge1_yuta_ktd/android/app/src/main/AndroidManifest.xml
@@ -15,6 +15,8 @@
                  the Android process has started. This theme is visible to the user
                  while the Flutter UI initializes. After that, this theme continues
                  to determine the Window background behind the Flutter UI. -->
+            <meta-data android:name="com.google.android.geo.API_KEY"
+                android:value="@string/MAP_API_KEY"/>
             <meta-data
               android:name="io.flutter.embedding.android.NormalTheme"
               android:resource="@style/NormalTheme"
