+++ Yuta-KTD/flutter_challenge1_yuta_ktd/ios/Runner/AppDelegate.swift
@@ -1,5 +1,7 @@
 import UIKit
 import Flutter
+import flutter_config
+import GoogleMaps
 
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate {
@@ -7,6 +9,7 @@ import Flutter
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
+    GMSServices.provideAPIKey(FlutterConfigPlugin.env(for: "MAP_API_KEY"))
     GeneratedPluginRegistrant.register(with: self)
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
