+++ Yuta-KTD/flutter_challenge1_yuta_ktd/lib/app.dart
@@ -1,0 +1,101 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      // home: const MapSample(),
+      home: const MyHomePage(title: 'GoogleMap呼び出すまでのテスト用'),
+    );
+  }
+}
+
+// TODO: GoogleMapを呼び出したいフェーズに移ったら削除
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: const Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'test',
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+// FIXME: これはサンプル
+// class MapSample extends StatefulWidget {
+//   const MapSample({Key? key}) : super(key: key);
+
+//   @override
+//   State<MapSample> createState() => MapSampleState();
+// }
+
+// class MapSampleState extends State<MapSample> {
+//   final Completer<GoogleMapController> _controller = Completer();
+
+//   static const CameraPosition _kGooglePlex = CameraPosition(
+//     target: LatLng(37.42796133580664, -122.085749655962),
+//     zoom: 14.4746,
+//   );
+
+//   static const CameraPosition _kLake = CameraPosition(
+//       bearing: 192.8334901395799,
+//       target: LatLng(37.43296265331129, -122.08832357078792),
+//       tilt: 59.440717697143555,
+//       zoom: 19.151926040649414);
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       body: GoogleMap(
+//         mapType: MapType.hybrid,
+//         initialCameraPosition: _kGooglePlex,
+//         onMapCreated: (GoogleMapController controller) {
+//           _controller.complete(controller);
+//         },
+//       ),
+//       floatingActionButton: FloatingActionButton.extended(
+//         onPressed: _goToTheLake,
+//         label: const Text('To the lake!'),
+//         icon: const Icon(Icons.directions_boat),
+//       ),
+//     );
+//   }
+
+//   Future<void> _goToTheLake() async {
+//     final GoogleMapController controller = await _controller.future;
+//     controller.animateCamera(CameraUpdate.newCameraPosition(_kLake));
+//   }
+// }
+
+
